add function to select slick based on dot point
